extract output folder from mto model when saving, so that we dont clutter up our model folder some more playing around with trip chaining analysis fix bug getting persod by a pumfId, the id needs to be a long sort the tour stops in reverse order, as that better represents real chronoligical order toString() functions for tour and visit for debugging

+++ src/main/java/de/tum/bgu/msm/dataAnalysis/surveyModel/mtoSurveyData.java
@@ -85,7 +85,7 @@ public class mtoSurveyData {
         return tripPurposes;
     }
 
-    public surveyPerson getPersonFromId(int id) {
+    public surveyPerson getPersonFromId(long id) {
         return personMap.get(id);
     }
